Move push to push.rs
Towards #290
